refactor: 마이페이지 메인 화면 내 sub view 추출
